requirements.txt
face-recognition
opencv-python
pillow
piexif
requests
beautifulsoup4
selenium
yandex-images-download
google-images-download
pandas
numpy
scikit-learn
nltk
spacy
transformers
torch
streamlit
plotly
matplotlib
seaborn
cryptography
pytest
unittest2
